libs/salome-cdk/pyproject.toml
[project]
name = "salome-cdk"
version = "1.0.0"
requires-python = ">=3.14"

[dependency-groups]
dev = [
    "aws-cdk-lib>=2.235.0",
    "constructs>=10.4.5",
]

[tool.ruff.lint]
ignore = [
    "F841", # unused-variable
]
